+++ modules/mitigations/__init__.py
@@ -6,19 +6,20 @@ def get_priority():
 
 def get_menu():
     return {
-        "name": "Mitigations", 
+        "display_name": mitigations_config.module_name, 
+        "module_name": mitigations_config.module_name,
         "url": "/mitigations/", 
         "external_link": False,
         "priority": mitigations_config.priority,
         "children": [
             {
-                "name": "Enterprise", 
+                "display_name": "Enterprise", 
                 "url": "/mitigations/enterprise/", 
                 "external_link": False,
                 "children": []
             },
             {
-                "name": "Mobile", 
+                "display_name": "Mobile", 
                 "url": "/mitigations/mobile/", 
                 "external_link": False,
                 "children": []