+++ modules/resources/resources.py
@@ -93,7 +93,7 @@ def check_menu_versions_module():
         remove from submenu 
     """
 
-    if not [key['name'] for key in modules.run_ptr if key['name'] == 'versions']:
+    if not [key['module_name'] for key in modules.run_ptr if key['module_name'] == 'versions']:
         util.buildhelpers.remove_element_from_sub_menu(resources_config.module_name, "Versions of ATT&CK")
 
 
@@ -125,7 +125,7 @@ def generate_working_with_attack():
         f"mobile-attack-{site_config.full_attack_version}",
         f"ics-attack-{site_config.full_attack_version}"
     ]
-    files_types = ["matrices", "mitigations", "relationships", "software", "groups", "tactics", "techniques"]
+    files_types = ["matrices", "mitigations", "relationships", "software", "groups", "tactics", "techniques", "datasources"]
 
     # Verify if directories exists
     if not os.path.isdir(site_config.web_directory):